+++ packages/terra-date-picker/src/DateInput.jsx
@@ -60,6 +60,15 @@ const defaultProps = {
   value: undefined,
 };
 
+const contextTypes = {
+  /* eslint-disable consistent-return */
+  intl: (context) => {
+    if (context.intl === undefined) {
+      return new Error('Please add locale prop to Base component to load translations');
+    }
+  },
+};
+
 // eslint-disable-next-line react/prefer-stateless-function
 class DatePickerInput extends React.Component {
   constructor(props) {
@@ -115,20 +124,20 @@ class DatePickerInput extends React.Component {
     // Since we want to show the picker only when the calendar button is clicked, we need to delete the onFocus handle that is passed in by react-datepicker.
     delete additionalInputProps.onFocus;
 
-    const dateValue = DateUtil.convertToISO8601(value, 'MM/DD/YYYY');
+    const dateValue = DateUtil.convertToISO8601(value, DateUtil.getFormatByLocale(this.context.intl.locale));
 
     return (
       (<div className={styles['custom-input']}>
         <input
           // Create a hidden input for storing the name and value attributes to use when submitting the form.
           // The data stored in the value attribute will be the visible date in the date input but in ISO 8601 format.
-          data-class="hidden-date-input"
+          data-terra-date-input-hidden
           type="hidden"
           name={name}
           value={dateValue}
         />
         <Input
-          {...additionalInputProps} // TODO: When forms is available, this.props.inputAttributes should be passed to the attrs props in the TextField component (attrs={this.props.inputAttributes}) instead of destructuring the inputAttributes prop here.
+          {...additionalInputProps}
           className={styles.input}
           type="text"
           name={'terra-date-'.concat(name)}
@@ -151,5 +160,6 @@ class DatePickerInput extends React.Component {
 
 DatePickerInput.propTypes = propTypes;
 DatePickerInput.defaultProps = defaultProps;
+DatePickerInput.contextTypes = contextTypes;
 
 export default DatePickerInput;