+++ packages/terra-date-picker/tests/nightwatch/components/DatePickerModalContainer.jsx
@@ -1,8 +1,11 @@
 import React from 'react';
+import Base from 'terra-base';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import AppDelegate from 'terra-app-delegate';
 import Button from 'terra-button';
 import DatePicker from '../../../lib/DatePicker';
+
+const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
 
 class ModalContainer extends React.Component {
   constructor(props) {
@@ -40,7 +43,13 @@ class ModalContainer extends React.Component {
         {props.app && props.app.releaseFocus ? <h4>Modal focus is released!</h4> : null }
         {props.app && props.app.requestFocus ? <h4>Modal focus is trapped!</h4> : null }
         <br />
-        <DatePicker name="date-picker-in-modal" releaseFocus={props.app.releaseFocus} requestFocus={props.app.requestFocus} />
+        <Base locale={locale}>
+          <DatePicker
+            name="date-picker-in-modal"
+            releaseFocus={props.app.releaseFocus}
+            requestFocus={props.app.requestFocus}
+          />
+        </Base>
         <br />
         <br />
         <Button className="close-disclosure" onClick={this.closeDisclosure}>Close Disclosure</Button>