+++ packages/terra-date-picker/tests/nightwatch/components/DatePickerStartDate.jsx
@@ -1,11 +1,16 @@
 import React from 'react';
+import Base from 'terra-base';
 import DatePicker from '../../../lib/DatePicker';
 
+const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
+
 const DatePickerStartDate = () => (
-  <DatePicker
-    name="date-input"
-    selectedDate={'2017-04-01'}
-  />
+  <Base locale={locale}>
+    <DatePicker
+      name="date-input"
+      selectedDate={'2017-04-01'}
+    />
+  </Base>
 );
 
 export default DatePickerStartDate;