+++ packages/terra-props-table/tests/nightwatch/MockPropsTable.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import PropsTable from '../../lib/PropsTable';
 
 // eslint-disable-next-line import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions
-import IconSrc from '!raw-loader!./MockComponent';
+import MockSrc from '!raw-loader!./MockComponent';
 
-const IconExamples = () => (
+const MockExample = () => (
   <div>
-    <PropsTable id="PropsTable" src={IconSrc} />
+    <PropsTable id="PropsTable" src={MockSrc} />
   </div>
 );
 
-export default IconExamples;
+export default MockExample;